+++ processor/src/test/resources/input/com/example/tracked/InheritProtectedAccessTrackedModel.java
@@ -1,0 +1,10 @@
+package com.example.tracked;
+
+import com.example.tracked.other.BaseModelProtectedAccess;
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent
+public class InheritProtectedAccessTrackedModel
+  extends BaseModelProtectedAccess
+{
+}