+++ processor/src/test/resources/input/com/example/tracked/other/BaseModelProtectedAccess.java
@@ -1,0 +1,17 @@
+package com.example.tracked.other;
+
+import org.realityforge.arez.annotations.OnDepsUpdated;
+import org.realityforge.arez.annotations.Track;
+
+public class BaseModelProtectedAccess
+{
+  @Track
+  protected void render( final long time, float someOtherParameter )
+  {
+  }
+
+  @OnDepsUpdated
+  protected final void onRenderDepsUpdated()
+  {
+  }
+}